finish compSddmmCoo()

+++ HiPC18/include/kernel.cuh
@@ -1,4 +1,6 @@
 #pragma once
 
-__global__ void compSddmmCoo(const size_t M, const size_t N, const size_t K,
-                             const float *matrixA, const float *matrixB, const float *matrixS, float *matrixP);
+__global__ void compSddmmCoo(const int M, const int N, const int K,
+                             const half *matrixA, const half *matrixB,
+                             const float *matrixS,
+                             float *matrixP);
